Add SoilCanopyModel diagnostics
This PR finalizes the addition of SoilCanopyModel diagnostics.
It also adds a macro to generate land compute functions,
And an optional argument for period averaging for default diagnostics.

+++ experiments/long_runs/land.jl
@@ -601,7 +601,7 @@ function setup_prob(t0, tf, Δt; outdir = outdir, nelements = (101, 15))
         land,
         t0;
         output_writer = nc_writer,
-        output_vars = :short,
+        output_vars = :long,
     )
 
     diagnostic_handler =
